requirements.txt
google_api_python_client==2.163.0
google_auth_oauthlib==1.2.1
protobuf==6.30.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
